Fixed up table too.  hopefully all kosher
Broke down and did the diagrams as static pictures hosted on imgur

Is there a way to overwrite an image on imgur?  That would be a
nice post commit-hook, if certain tex files were overwritten

Have it run tex on the diagrams, convert to png, flatten, and
upload to that imgur address.  Something to consider.